Move base graphql files into subdir
